requirements.txt
python-dotenv==1.0.1
huggingface-hub==0.24.7
accelerate>=0.16.0,<1
transformers==4.44.2
safetensors==0.4.5
langchain==0.3.1
langchain-community==0.3.0
langchain-experimental==0.3.1
tiktoken==0.7.0
sentence-transformers==3.1.0
langchain-huggingface==0.1.0
sentencepiece==0.2.0
pypdf==4.3.1
langchain-groq==0.2.0
groq==0.11.0
streamlit
pandas==2.2.2

